+++ src/cli/connect.rs
@@ -208,14 +208,16 @@ fn sql_repl(args: args::Connect) -> Result<(), ReplError> {
 
     const HELP_MESSAGE: &'static str = "
     Available backslash commands:
-        \\e            Open the editor specified by the EDITOR environment variable
-        \\help         Show this screen
+        \\e                              Open the editor specified by the EDITOR environment variable
+        \\help                           Show this screen
+        \\set delimiter shiny-delimiter  Set console delimiter to 'shiny-delimiter'
+        \\set delimiter default          Reset console delimiter to default (enter)
 
     Available hotkeys:
-        Enter         Submit the request
-        Alt  + Enter  Insert a newline character
-        Ctrl + C      Discard current input
-        Ctrl + D      Quit interactive console";
+        Enter                           Submit the request
+        Alt  + Enter                    Insert a newline character
+        Ctrl + C                        Discard current input
+        Ctrl + D                        Quit interactive console";
 
     while let Some(command) = console.read()? {
         match command {